backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
pydantic-settings==2.5.0
anthropic==0.34.0
python-multipart==0.0.12
pillow==10.4.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
